Revert "Add more test cases"
