+++ package.json
@@ -4,14 +4,11 @@
   "private": true,
   "dependencies": {
     "@blocto/fcl": "^0.0.78-alpha.9",
-    "@blocto/sdk": "^0.2.21",
     "@chakra-ui/icons": "^1.1.1",
     "@chakra-ui/react": "^1.7.4",
     "@emotion/react": "^11",
     "@emotion/styled": "^11",
     "@onflow/types": "^0.0.6",
-    "@solana/buffer-layout": "^4.0.0",
-    "@solana/web3.js": "^1.36.0",
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
@@ -26,8 +23,7 @@
     "react-scripts": "4.0.3",
     "sha3": "^2.1.4",
     "typescript": "^4.5.4",
-    "web-vitals": "^2.1.3",
-    "web3": "^1.7.1"
+    "web-vitals": "^2.1.3"
   },
   "scripts": {
     "start": "react-scripts start",
