+++ src/App.tsx
@@ -2,17 +2,13 @@ import React from "react";
 import { ChakraProvider } from "@chakra-ui/react";
 import { ReactJSXElement } from "@emotion/react/types/jsx-namespace";
 import Header from "./components/Header";
-import Playground from "./components/Playground";
-import ContextProvider from "./context/Context";
+import Editor from "./components/Editor";
 
-const App = (): ReactJSXElement => {
-  return (
-    <ChakraProvider>
-      <ContextProvider>
-        <Header />
-        <Playground />
-      </ContextProvider>
-    </ChakraProvider>
-  );
-};
+const App = (): ReactJSXElement => (
+  <ChakraProvider>
+    <Header />
+    <Editor />
+  </ChakraProvider>
+);
+
 export default App;
