--- src/components/EvmChainSelect.tsx
@@ -1,36 +1,0 @@
-import React, { useContext } from "react";
-import { Button, Menu, MenuButton, MenuList, MenuItem } from "@chakra-ui/react";
-import { ChevronDownIcon } from "@chakra-ui/icons";
-import { startCase } from "lodash";
-import { ReactJSXElement } from "@emotion/react/types/jsx-namespace";
-import { Context } from "../context/Context";
-import { EvmChain } from "../types/ChainTypes";
-
-const EvmChainSelect: React.FC = ({}): ReactJSXElement => {
-  const { chain, switchChain } = useContext(Context);
-  return (
-    <Menu>
-      <MenuButton as={Button} rightIcon={<ChevronDownIcon />} width="130px">
-        {startCase(chain)}
-      </MenuButton>
-      <MenuList>
-        {Object.values(EvmChain).map((chain) => (
-          <MenuItem
-            key={chain}
-            pl={5}
-            color="gray.700"
-            onClick={() => {
-              if (switchChain) {
-                switchChain(chain);
-              }
-            }}
-          >
-            {startCase(chain)}
-          </MenuItem>
-        ))}
-      </MenuList>
-    </Menu>
-  );
-};
-
-export default EvmChainSelect;
