--- src/contracts/fuji.ts
@@ -1,69 +1,0 @@
-import { AbiItem } from "web3-utils";
-
-const contractAddress = "0x247954ec015C4E641c99C47Cb8766b35cdb1Eb53";
-const contractAbi: AbiItem[] = [
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: false,
-        internalType: "uint256",
-        name: "value",
-        type: "uint256",
-      },
-    ],
-    name: "updateValue",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: false,
-        internalType: "uint256",
-        name: "value",
-        type: "uint256",
-      },
-    ],
-    name: "updateValue2",
-    type: "event",
-  },
-  {
-    constant: false,
-    inputs: [{ internalType: "uint256", name: "newValue", type: "uint256" }],
-    name: "setValue",
-    outputs: [],
-    payable: false,
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    constant: false,
-    inputs: [{ internalType: "uint256", name: "newValue", type: "uint256" }],
-    name: "setValue2",
-    outputs: [],
-    payable: false,
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    constant: true,
-    inputs: [],
-    name: "value",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    payable: false,
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    constant: true,
-    inputs: [],
-    name: "value2",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    payable: false,
-    stateMutability: "view",
-    type: "function",
-  },
-];
-
-export { contractAddress, contractAbi };
