--- src/contracts/solanaDevnet.ts
@@ -1,5 +1,0 @@
-const programId = "G4YkbRN4nFQGEUg4SXzPsrManWzuk8bNq9JaMhXepnZ6";
-
-const accountPubKey = "4AXy5YYCXpMapaVuzKkz25kVHzrdLDgKN3TiQvtf1Eu8";
-
-export { programId, accountPubKey };
