--- src/scripts/flow/SignMessage.ts
@@ -1,7 +1,0 @@
-import ScriptTypes from "../../types/ScriptTypes";
-
-export const signMessage = {
-  type: ScriptTypes.SIGN,
-  script: "",
-  args: [{ type: "String", comment: "message" }],
-};
