--- src/types/ChainTypes.ts
@@ -1,15 +1,0 @@
-export enum OtherChain {
-  Flow = "flow",
-  Solana = "solana",
-}
-
-export enum EvmChain {
-  Ethereum = "ethereum",
-  Bsc = "bsc",
-  Polygon = "polygon",
-  Avalanche = "avalance",
-}
-
-export type ChainsType = OtherChain | EvmChain;
-
-export const Chains = { ...OtherChain, ...EvmChain };
