+++ src/types/ScriptTypes.ts
@@ -1,8 +1,6 @@
 enum ScriptTypes {
   SCRIPT,
   TX,
-  SIGN,
-  CONTRACT,
 }
 
 export default ScriptTypes;
