--- src/utils/base64ToUint8Array.ts
@@ -1,3 +1,0 @@
-export const base64ToUint8Array = (base64: string): Uint8Array => {
-  return Uint8Array.from(window.atob(base64), (c) => c.charCodeAt(0));
-};
