--- src/utils/formatAddress.ts
@@ -1,8 +1,0 @@
-export const formatAddress = (address: string | undefined): string => {
-  if (!address) {
-    return "";
-  }
-  const splitted = address.split("");
-  splitted.splice(6, address.length - 10, "...");
-  return splitted.join("");
-};
